+++ Craftsman/Builders/SolutionBuilder.cs
@@ -66,6 +66,7 @@ public class SolutionBuilder
         new ProgramBuilder(_utilities).CreateWebApiProgram(srcDirectory, useJwtAuth, projectBaseName, useCustomErrorHandler);
         new ServiceConfigurationBuilder(_utilities).CreateWebAppServiceConfiguration(srcDirectory, projectBaseName, useCustomErrorHandler);
         new ConstsResourceBuilder(_utilities).CreateLocalConfig(srcDirectory, projectBaseName);
+        new QueryKitConfigBuilder(_utilities).CreateConfig(srcDirectory, projectBaseName);
         new LoggingConfigurationBuilder(_utilities).CreateWebApiConfigFile(srcDirectory, projectBaseName);
         new InfrastructureServiceRegistrationBuilder(_utilities).CreateInfrastructureServiceExtension(srcDirectory, projectBaseName);